Feat/ai translate result

+++ ai-server/app/schemas/diagnosis.py
@@ -69,7 +69,7 @@ class CVResult(BaseModel):
 class DiagnosisResult(BaseModel):
     id: int
     disease: str
-    probability: float
+    probability: int
     reason: str
